+++ R/tilt_profile.R
@@ -17,33 +17,33 @@
 #'
 #' class(profile(data))
 #'
-#' is_profile(profile(data))
-profile <- function(data) {
-  validate_profile(data)
-  new_profile(data)
+#' is_tilt_profile(profile(data))
+tilt_profile <- function(data) {
+  validate_tilt_profile(data)
+  new_tilt_profile(data)
 }
 
-validate_profile <- function(data) {
-  if (!has_profile_names(data)) {
+validate_tilt_profile <- function(data) {
+  if (!has_tilt_profile_names(data)) {
     msg <- "`data` must have columns `companies_id`, `product`, and `company`."
-    abort(msg, class = "validate_profile_top_level_names")
+    abort(msg, class = "validate_tilt_profile_top_level_names")
   }
 
   if (!is.list(data$product) | !is.list(data$company)) {
     msg <- "The `product` and `company` columns must be a list."
-    abort(msg, class = "validate_profile_type")
+    abort(msg, class = "validate_tilt_profile_type")
   }
 
   data
 }
 
-has_profile_names <- function(data) {
+has_tilt_profile_names <- function(data) {
   identical(names(data), c("companies_id", "product", "company"))
 }
 
-new_profile <- function(data) {
+new_tilt_profile <- function(data) {
   stopifnot(is.data.frame(data))
-  structure(data, class = c("profile", class(data)))
+  structure(data, class = c("tilt_profile", class(data)))
 }
 
 document_profile <- function() {
@@ -55,6 +55,6 @@ document_profile <- function() {
 
 #' @rdname profile
 #' @export
-is_profile <- function(data) {
-  inherits(data, "profile")
+is_tilt_profile <- function(data) {
+  inherits(data, "tilt_profile")
 }